Move JSON conversion functions to a separate class
