+++ panc/src/main/java/org/quattor/pan/output/JsonFormatter.java
@@ -21,28 +21,9 @@
 package org.quattor.pan.output;
 
 import java.io.PrintWriter;
-import java.lang.reflect.Type;
 
-import org.quattor.pan.dml.data.BooleanProperty;
-import org.quattor.pan.dml.data.DoubleProperty;
-import org.quattor.pan.dml.data.HashResource;
-import org.quattor.pan.dml.data.ListResource;
-import org.quattor.pan.dml.data.LongProperty;
-import org.quattor.pan.dml.data.Property;
-import org.quattor.pan.dml.data.ProtectedHashResource;
-import org.quattor.pan.dml.data.ProtectedListResource;
-import org.quattor.pan.dml.data.Resource;
-import org.quattor.pan.dml.data.StringProperty;
 import org.quattor.pan.tasks.FinalResult;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonPrimitive;
-import com.google.gson.JsonSerializationContext;
-import com.google.gson.JsonSerializer;
+import org.quattor.pan.utils.JsonUtils;
 
 public class JsonFormatter extends AbstractFormatter {
 
@@ -61,75 +42,9 @@ public class JsonFormatter extends AbstractFormatter {
 	}
 
 	protected void write(FinalResult result, PrintWriter ps) throws Exception {
-
-		Gson gson = new GsonBuilder()
-				.registerTypeAdapter(BooleanProperty.class,
-						new PropertySerializer())
-				.registerTypeAdapter(DoubleProperty.class,
-						new PropertySerializer())
-				.registerTypeAdapter(LongProperty.class,
-						new PropertySerializer())
-				.registerTypeAdapter(StringProperty.class,
-						new PropertySerializer())
-				.registerTypeAdapter(HashResource.class, new HashSerializer())
-				.registerTypeAdapter(ProtectedHashResource.class,
-						new HashSerializer())
-				.registerTypeAdapter(ListResource.class, new ListSerializer())
-				.registerTypeAdapter(ProtectedListResource.class,
-						new ListSerializer()).setPrettyPrinting().create();
-
-		ps.write(gson.toJson(result.getRoot()));
+		JsonUtils json = JsonUtils.getInstance();
+		ps.write(json.toJson(result.getRoot()));
 		ps.close();
-
-	}
-
-	private static class PropertySerializer implements JsonSerializer<Property> {
-		public JsonElement serialize(Property src, Type typeOfSrc,
-				JsonSerializationContext context) {
-
-			if (src instanceof BooleanProperty) {
-				return new JsonPrimitive(((BooleanProperty) src).getValue());
-			} else if (src instanceof DoubleProperty) {
-				return new JsonPrimitive(((DoubleProperty) src).getValue());
-			} else if (src instanceof LongProperty) {
-				return new JsonPrimitive(((LongProperty) src).getValue());
-			} else {
-				return new JsonPrimitive(src.getValue().toString());
-			}
-		}
-	}
-
-	private static class HashSerializer implements JsonSerializer<HashResource> {
-
-		public JsonElement serialize(HashResource src, Type typeOfSrc,
-				JsonSerializationContext context) {
-
-			JsonObject map = new JsonObject();
-
-			for (Resource.Entry entry : src) {
-				String property = entry.getKey().toString();
-				JsonElement value = context.serialize(entry.getValue());
-				map.add(property, value);
-			}
-
-			return map;
-		}
-	}
-
-	private static class ListSerializer implements JsonSerializer<ListResource> {
-
-		public JsonElement serialize(ListResource src, Type typeOfSrc,
-				JsonSerializationContext context) {
-
-			JsonArray array = new JsonArray();
-
-			for (Resource.Entry entry : src) {
-				JsonElement value = context.serialize(entry.getValue());
-				array.add(value);
-			}
-
-			return array;
-		}
 	}
 
 }
